refacto: move TargetActionsHistory to a new file
